pyproject.toml
[project]
name = "invest-harness"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "jsonschema>=4.20.0",
    "python-dotenv>=1.0.0",
    "chromadb>=0.4.0",
    "tushare>=1.4.0",
    "yfinance>=0.2.30",
    "requests>=2.31.0",
    "PyYAML>=6.0",
]

[project.scripts]
invest-harness = "scripts.harness_cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.12.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
